refactor: 상태들을 패키지 단위로 묶음
