+++ src/main/java/chess/domain/game/state/finished/WhiteWin.java
@@ -1,5 +1,6 @@
-package chess.domain.game;
+package chess.domain.game.state.finished;
 
+import chess.domain.game.ChessGame;
 import chess.domain.piece.white.WhitePiece;
 
 import java.util.Optional;
