+++ src/main/java/chess/domain/game/state/idle/Idle.java
@@ -1,4 +1,7 @@
-package chess.domain.game;
+package chess.domain.game.state.idle;
+
+import chess.domain.game.ChessGame;
+import chess.domain.game.state.Started;
 
 public abstract class Idle extends Started {
