+++ src/main/java/chess/domain/game/state/running/BlackTurn.java
@@ -1,5 +1,8 @@
-package chess.domain.game;
+package chess.domain.game.state.running;
 
+import chess.domain.game.state.finished.BlackWin;
+import chess.domain.game.ChessGame;
+import chess.domain.game.state.finished.End;
 import chess.domain.piece.Position;
 
 public class BlackTurn extends Running {
