+++ src/test/java/chess/domain/game/ReadyTest.java
@@ -1,5 +1,6 @@
 package chess.domain.game;
 
+import chess.domain.game.state.idle.Ready;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
