+++ src/test/java/chess/domain/game/WhiteTurnTest.java
@@ -1,6 +1,7 @@
 package chess.domain.game;
 
 import chess.domain.board.Board;
+import chess.domain.game.state.running.WhiteTurn;
 import chess.domain.piece.PieceFactory;
 import chess.domain.piece.Position;
 import org.junit.jupiter.api.BeforeEach;
